+++ example/lib/pages/cache_page.dart
@@ -1,0 +1,57 @@
+import 'package:better_player/better_player.dart';
+import 'package:better_player_example/constants.dart';
+import 'package:flutter/material.dart';
+
+class CachePage extends StatefulWidget {
+  @override
+  _CachePageState createState() => _CachePageState();
+}
+
+class _CachePageState extends State<CachePage> {
+  BetterPlayerController _betterPlayerController;
+
+  @override
+  void initState() {
+    BetterPlayerConfiguration betterPlayerConfiguration =
+        BetterPlayerConfiguration(
+      aspectRatio: 16 / 9,
+      fit: BoxFit.contain,
+    );
+    BetterPlayerDataSource dataSource = BetterPlayerDataSource(
+      BetterPlayerDataSourceType.NETWORK,
+      Constants.forBiggerBlazesUrl,
+      cacheConfiguration: BetterPlayerCacheConfiguration(useCache: true),
+    );
+    _betterPlayerController = BetterPlayerController(betterPlayerConfiguration);
+    _betterPlayerController.setupDataSource(dataSource);
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Normal player"),
+      ),
+      body: Column(
+        children: [
+          const SizedBox(height: 8),
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 16),
+            child: Text(
+              "Player with cache enabled. To test this feature, first plays "
+              "video, then leave this page, turn internet off and enter "
+              "page again. You should be able to play video without "
+              "internet connection.",
+              style: TextStyle(fontSize: 16),
+            ),
+          ),
+          AspectRatio(
+            aspectRatio: 16 / 9,
+            child: BetterPlayer(controller: _betterPlayerController),
+          ),
+        ],
+      ),
+    );
+  }
+}