+++ example/lib/pages/event_listener_page.dart
@@ -1,0 +1,91 @@
+import 'dart:async';
+
+import 'package:better_player/better_player.dart';
+import 'package:better_player_example/constants.dart';
+import 'package:flutter/material.dart';
+
+class EventListenerPage extends StatefulWidget {
+  @override
+  _EventListenerPageState createState() => _EventListenerPageState();
+}
+
+class _EventListenerPageState extends State<EventListenerPage> {
+  BetterPlayerController _betterPlayerController;
+  List<BetterPlayerEvent> events = List();
+  StreamController<DateTime> _eventStreamController =
+      StreamController.broadcast();
+
+  @override
+  void dispose() {
+    _eventStreamController.close();
+    super.dispose();
+  }
+
+  @override
+  void initState() {
+    BetterPlayerConfiguration betterPlayerConfiguration =
+        BetterPlayerConfiguration(
+      aspectRatio: 16 / 9,
+      fit: BoxFit.contain,
+    );
+    BetterPlayerDataSource dataSource = BetterPlayerDataSource(
+        BetterPlayerDataSourceType.NETWORK, Constants.forBiggerBlazesUrl);
+    _betterPlayerController = BetterPlayerController(betterPlayerConfiguration);
+    _betterPlayerController.setupDataSource(dataSource);
+    _betterPlayerController.addEventsListener((event) {
+      events.insert(0, event);
+
+      ///Used to refresh only list of events
+      _eventStreamController.add(DateTime.now());
+    });
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Normal player"),
+      ),
+      body: Column(
+        children: [
+          const SizedBox(height: 8),
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 16),
+            child: Text(
+              "Better Player exposes events which can be listened with event listener. Start player to see events flowing.",
+              style: TextStyle(fontSize: 16),
+            ),
+          ),
+          const SizedBox(height: 8),
+          AspectRatio(
+            aspectRatio: 16 / 9,
+            child: BetterPlayer(controller: _betterPlayerController),
+          ),
+          const SizedBox(height: 8),
+          Expanded(
+            child: StreamBuilder(
+              stream: _eventStreamController.stream,
+              builder: (context, snapshot) {
+                return ListView(
+                  children: events
+                      .map(
+                        (event) => Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: [
+                            Text(
+                                "Event: ${event.betterPlayerEventType} parameters: ${(event.parameters ?? {}).toString()}"),
+                            Divider(),
+                          ],
+                        ),
+                      )
+                      .toList(),
+                );
+              },
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}