+++ example/lib/utils.dart
@@ -1,0 +1,8 @@
+import 'package:path_provider/path_provider.dart';
+
+class Utils {
+  static Future<String> getFileUrl(String fileName) async {
+    final directory = await getApplicationDocumentsDirectory();
+    return "${directory.path}/$fileName";
+  }
+}